auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "auth service"
authors = [
    {name = "Az1muth",email = "yxyxy@mail.ru"}
]
license = {text = "MIT"}
# readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]

[[project.packages]]
include = "auth"

[[project.packages]]
include = "users"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
